Revert "temp"
This was not meant to be here.

This reverts commit deb02c44eb0b4f159fbbd012c36ab7f6b020a9d2.

+++ external/Box2D/Collision/Shapes/b2PolygonShape.cpp
@@ -27,7 +27,7 @@ b2Shape* b2PolygonShape::Clone(b2BlockAllocator* allocator) const
 	return clone;
 }
 
-void b2PolygonShape::SetAsBox(qreal hx, float32 hy)
+void b2PolygonShape::SetAsBox(float32 hx, float32 hy)
 {
 	m_count = 4;
 	m_vertices[0].Set(-hx, -hy);
@@ -41,7 +41,7 @@ void b2PolygonShape::SetAsBox(qreal hx, float32 hy)
 	m_centroid.SetZero();
 }
 
-void b2PolygonShape::SetAsBox(qreal hx, float32 hy, const b2Vec2& center, float32 angle)
+void b2PolygonShape::SetAsBox(float32 hx, float32 hy, const b2Vec2& center, float32 angle)
 {
 	m_count = 4;
 	m_vertices[0].Set(-hx, -hy);
@@ -76,7 +76,7 @@ static b2Vec2 ComputeCentroid(const b2Vec2* vs, int32 count)
 	b2Assert(count >= 3);
 
 	b2Vec2 c; c.Set(0.0f, 0.0f);
-	qreal area = 0.0f;
+	float32 area = 0.0f;
 
 	// pRef is the reference point for forming triangles.
 	// It's location doesn't change the result (except for rounding error).
@@ -90,7 +90,7 @@ static b2Vec2 ComputeCentroid(const b2Vec2* vs, int32 count)
 	pRef *= 1.0f / count;
 #endif
 
-	const qreal inv3 = 1.0f / 3.0f;
+	const float32 inv3 = 1.0f / 3.0f;
 
 	for (int32 i = 0; i < count; ++i)
 	{
@@ -102,9 +102,9 @@ static b2Vec2 ComputeCentroid(const b2Vec2* vs, int32 count)
 		b2Vec2 e1 = p2 - p1;
 		b2Vec2 e2 = p3 - p1;
 
-		qreal D = b2Cross(e1, e2);
+		float32 D = b2Cross(e1, e2);
 
-		qreal triangleArea = 0.5f * D;
+		float32 triangleArea = 0.5f * D;
 		area += triangleArea;
 
 		// Area weighted centroid
@@ -165,10 +165,10 @@ void b2PolygonShape::Set(const b2Vec2* vertices, int32 count)
 
 	// Find the right most point on the hull
 	int32 i0 = 0;
-	qreal x0 = ps[0].x;
+	float32 x0 = ps[0].x;
 	for (int32 i = 1; i < n; ++i)
 	{
-		qreal x = ps[i].x;
+		float32 x = ps[i].x;
 		if (x > x0 || (x == x0 && ps[i].y < ps[i0].y))
 		{
 			i0 = i;
@@ -195,7 +195,7 @@ void b2PolygonShape::Set(const b2Vec2* vertices, int32 count)
 
 			b2Vec2 r = ps[ie] - ps[hull[m]];
 			b2Vec2 v = ps[j] - ps[hull[m]];
-			qreal c = b2Cross(r, v);
+			float32 c = b2Cross(r, v);
 			if (c < 0.0f)
 			{
 				ie = j;
@@ -254,7 +254,7 @@ bool b2PolygonShape::TestPoint(const b2Transform& xf, const b2Vec2& p) const
 
 	for (int32 i = 0; i < m_count; ++i)
 	{
-		qreal dot = b2Dot(m_normals[i], pLocal - m_vertices[i]);
+		float32 dot = b2Dot(m_normals[i], pLocal - m_vertices[i]);
 		if (dot > 0.0f)
 		{
 			return false;
@@ -274,7 +274,7 @@ bool b2PolygonShape::RayCast(b2RayCastOutput* output, const b2RayCastInput& inpu
 	b2Vec2 p2 = b2MulT(xf.q, input.p2 - xf.p);
 	b2Vec2 d = p2 - p1;
 
-	qreal lower = 0.0f, upper = input.maxFraction;
+	float32 lower = 0.0f, upper = input.maxFraction;
 
 	int32 index = -1;
 
@@ -283,8 +283,8 @@ bool b2PolygonShape::RayCast(b2RayCastOutput* output, const b2RayCastInput& inpu
 		// p = p1 + a * d
 		// dot(normal, p - v) = 0
 		// dot(normal, p1 - v) + a * dot(normal, d) = 0
-		qreal numerator = b2Dot(m_normals[i], m_vertices[i] - p1);
-		qreal denominator = b2Dot(m_normals[i], d);
+		float32 numerator = b2Dot(m_normals[i], m_vertices[i] - p1);
+		float32 denominator = b2Dot(m_normals[i], d);
 
 		if (denominator == 0.0f)
 		{	
@@ -355,7 +355,7 @@ void b2PolygonShape::ComputeAABB(b2AABB* aabb, const b2Transform& xf, int32 chil
 	aabb->upperBound = upper + r;
 }
 
-void b2PolygonShape::ComputeMass(b2MassData* massData, qreal density) const
+void b2PolygonShape::ComputeMass(b2MassData* massData, float32 density) const
 {
 	// Polygon mass, centroid, and inertia.
 	// Let rho be the polygon density in mass per unit area.
@@ -384,8 +384,8 @@ void b2PolygonShape::ComputeMass(b2MassData* massData, qreal density) const
 	b2Assert(m_count >= 3);
 
 	b2Vec2 center; center.Set(0.0f, 0.0f);
-	qreal area = 0.0f;
-	qreal I = 0.0f;
+	float32 area = 0.0f;
+	float32 I = 0.0f;
 
 	// s is the reference point for forming triangles.
 	// It's location doesn't change the result (except for rounding error).
@@ -398,7 +398,7 @@ void b2PolygonShape::ComputeMass(b2MassData* massData, qreal density) const
 	}
 	s *= 1.0f / m_count;
 
-	const qreal k_inv3 = 1.0f / 3.0f;
+	const float32 k_inv3 = 1.0f / 3.0f;
 
 	for (int32 i = 0; i < m_count; ++i)
 	{
@@ -406,19 +406,19 @@ void b2PolygonShape::ComputeMass(b2MassData* massData, qreal density) const
 		b2Vec2 e1 = m_vertices[i] - s;
 		b2Vec2 e2 = i + 1 < m_count ? m_vertices[i+1] - s : m_vertices[0] - s;
 
-		qreal D = b2Cross(e1, e2);
+		float32 D = b2Cross(e1, e2);
 
-		qreal triangleArea = 0.5f * D;
+		float32 triangleArea = 0.5f * D;
 		area += triangleArea;
 
 		// Area weighted centroid
 		center += triangleArea * k_inv3 * (e1 + e2);
 
-		qreal ex1 = e1.x, ey1 = e1.y;
-		qreal ex2 = e2.x, ey2 = e2.y;
+		float32 ex1 = e1.x, ey1 = e1.y;
+		float32 ex2 = e2.x, ey2 = e2.y;
 
-		qreal intx2 = ex1*ex1 + ex2*ex1 + ex2*ex2;
-		qreal inty2 = ey1*ey1 + ey2*ey1 + ey2*ey2;
+		float32 intx2 = ex1*ex1 + ex2*ex1 + ex2*ex2;
+		float32 inty2 = ey1*ey1 + ey2*ey1 + ey2*ey2;
 
 		I += (0.25f * k_inv3 * D) * (intx2 + inty2);
 	}
@@ -455,7 +455,7 @@ bool b2PolygonShape::Validate() const
 			}
 
 			b2Vec2 v = m_vertices[j] - p;
-			qreal c = b2Cross(e, v);
+			float32 c = b2Cross(e, v);
 			if (c < 0.0f)
 			{
 				return false;
