+++ external/Box2D/Collision/b2TimeOfImpact.h
@@ -29,7 +29,7 @@ struct b2TOIInput
 	b2DistanceProxy proxyB;
 	b2Sweep sweepA;
 	b2Sweep sweepB;
-	qreal tMax;		// defines sweep interval [0, tMax]
+	float32 tMax;		// defines sweep interval [0, tMax]
 };
 
 // Output parameters for b2TimeOfImpact.
@@ -45,7 +45,7 @@ struct b2TOIOutput
 	};
 
 	State state;
-	qreal t;
+	float32 t;
 };
 
 /// Compute the upper bound on time before two shapes penetrate. Time is represented as
